Split HopsanTypes into two separate files
HString and HVector (HVector is header only)
